feat(RebuyDialog): move XButton to own component, change PropTypes for ConfirmDialog
